+++ BraintreeDataCollector/src/androidTest/java/com/braintreepayments/api/DataCollectorTest.java
@@ -54,7 +54,7 @@ public class DataCollectorTest {
 
         SharedPreferencesHelper.overrideConfigurationCache(mActivity, authorization, configuration);
 
-        BraintreeClient braintreeClient = new BraintreeClient(authorization, mActivity);
+        BraintreeClient braintreeClient = new BraintreeClient(mActivity, authorization);
         DataCollector sut = new DataCollector(braintreeClient);
 
         sut.collectDeviceData(mActivity, new DataCollectorCallback() {
@@ -86,7 +86,7 @@ public class DataCollectorTest {
         Authorization authorization = Authorization.fromString(Fixtures.TOKENIZATION_KEY);
         SharedPreferencesHelper.overrideConfigurationCache(mActivity, authorization, configuration);
 
-        BraintreeClient braintreeClient = new BraintreeClient(authorization, mActivity);
+        BraintreeClient braintreeClient = new BraintreeClient(mActivity, authorization);
         DataCollector sut = new DataCollector(braintreeClient);
 
         sut.collectDeviceData(mActivity, "600001", new DataCollectorCallback() {
@@ -115,7 +115,7 @@ public class DataCollectorTest {
 
         SharedPreferencesHelper.overrideConfigurationCache(mActivity, authorization, configuration);
 
-        BraintreeClient braintreeClient = new BraintreeClient(authorization, mActivity);
+        BraintreeClient braintreeClient = new BraintreeClient(mActivity, authorization);
         DataCollector sut = new DataCollector(braintreeClient);
 
         sut.collectDeviceData(mActivity, new DataCollectorCallback() {